raiko-host compile
