+++ raiko-host/Cargo.toml
@@ -9,6 +9,8 @@ path = "src/util/lib.rs"
 
 [dependencies]
 anyhow = "1.0"
+alloy-sol-types = { version = "0.4", optional = true }
+alloy-primitives = { version = "0.4", default-features = false }
 bincode = "1.3.3"
 bonsai-sdk = { workspace = true }
 bytemuck = "1.13"
@@ -16,6 +18,8 @@ clap = { version = "4.0", features = ["derive"] }
 structopt = "0.3.26"
 structopt-toml = "0.5.1"
 ethers-core = { version = "2.0" }
+ethers-providers = { version = "2.0" }
+flate2 = "1.0.28"
 hex = "0.4.3"
 revm = { workspace = true }
 serde = "1.0"
@@ -26,6 +30,7 @@ zeth-lib = { path = "../lib", features = ["taiko", "server"] }
 zeth-primitives = { path = "../primitives", features = ["taiko"] }
 serde_json = "1.0"
 hyper = { version = "0.14.27", features = ["server"] }
+hashbrown = { workspace = true }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
 tracing-appender = "0.2"
@@ -34,6 +39,7 @@ lru_time_cache = "0.11.11"
 prometheus = "0.13.3"
 lazy_static = "1.4.0"
 once_cell = "1.8.0"
+thiserror = "1.0"
 
 [dev-dependencies]
 assert_cmd = "2.0"
@@ -41,4 +47,4 @@ rstest = "0.18"
 
 [features]
 default = ["taiko"]
-taiko = []
+taiko = ["alloy-sol-types"]
