+++ raiko-host/src/prover/error.rs
@@ -1,13 +1,15 @@
 use std::fmt;
+use thiserror::Error as ThisError;
 
-#[derive(Debug)]
+#[derive(ThisError, Debug)]
 pub enum Error {
     Io(std::io::Error),
-    Anyhow(anyhow::Error),
+    Anyhow(#[from] anyhow::Error),
     Serde(serde_json::Error),
     JoinHandle(tokio::task::JoinError),
     String(String),
 }
+
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -24,12 +26,6 @@ impl fmt::Display for Error {
 impl From<std::io::Error> for Error {
     fn from(e: std::io::Error) -> Self {
         Error::Io(e)
-    }
-}
-
-impl From<anyhow::Error> for Error {
-    fn from(e: anyhow::Error) -> Self {
-        Error::Anyhow(e)
     }
 }
