+++ raiko-host/src/prover/mod.rs
@@ -8,3 +8,4 @@ pub mod proof;
 pub mod request;
 pub mod server;
 pub mod utils;
+pub mod taiko_extra;
