+++ raiko-host/src/util/provider/file_provider.rs
@@ -16,6 +16,7 @@ use std::{
     collections::{BTreeMap, HashMap},
     fs::File,
     io::{Read, Write},
+    path::{Path, PathBuf}
 };
 
 use anyhow::{anyhow, Result};
